requirements.txt
fastapi
uvicorn
pandas
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
tabulate
langchain[ollama]